# specification/api.yml
swagger: "2.0"
info:
  title: NicoNihonGeemu API
  description: A REST API for managing NicoNihonGeemu decks and access dictionnary definitions
  version: 0.1.0
paths:
...
  /deck:
    get:
      summary: Retrieves the user's decks
      operationId: fetchDecks
      tags:
        - deck
      responses:
        '200':
          description: An array of decks
    post:
      summary: Creates a new deck
      operationId: createDeck
      tags:
        - deck
      schema:
        type: object
        properties:
          name:
            type: string
      responses:
        '200':
          description: 'Deck creation successful'
        '409':
          description: 'Already existing deck'
        '400':
          description: 'Invalid name'
  /deck/{deckId}:
    get:
      summary: Retrieves the deck corresponding to the given id
      operationId: fetchDeck
      parameters:
      - in: path
        name: deckId
        type: string
        description: The id of the deck to retrieve
      tags:
        - deck
      responses:
        '200':
          description: The deck corresponding to the given id
          schema:
            $ref: '#/definitions/Deck'
        '403':
          description: Not allowed to access this deck
        '404':
          description: Deck not found
    put:
      summary: Modifies a deck's metadata
      operationId: modifyDeck
      tags:
        - deck
      parameters:
      - in: path
        name: deckId
        type: string
        description: The id of the deck to retrieve
      - in: body
        name: deckMetadata
        required: true
        schema:
          $ref: '#/definitions/DeckMetadata'
      responses:
        '200':
          description: Modifies a deck's metadata
        '403':
          description: Not allowed to access this deck
        '404':
          description: Deck not found
    delete:
      summary: Deletes the deck with the given id
      tags:
        - deck
      parameters:
      - in: path
        name: deckId
        type: string
        description: The id of the deck to retrieve
      responses:
        '200':
          description: Deck successfully deleted
        '403':
          description: Not allowed to delete this deck
        '404':
          description: Deck not found
...
definitions:
...
  Deck:
    properties:
      name:
        type: string
      cardList:
        type: array
        items:
          $ref: '#/definitions/Card'
  Card:
    properties:
      front:
        type: string
      back:
        type: string